refactored calendarView out of scheduleVC
