+++ travel-weather/Views/DateCell.swift
@@ -49,7 +49,7 @@ class DateCell: JTAppleCell {
             selectedTrailingConstraint
         ])
         
-        selectedView.layer.cornerRadius = (contentView.frame.size.width - 20) / 2.0
+        selectedView.layer.cornerRadius = (contentView.frame.size.height - 20) / 2.0
         selectedView.backgroundColor = UIColor.white.withAlphaComponent(0.5)
         
     }
@@ -78,6 +78,7 @@ class DateCell: JTAppleCell {
         
     }
     func selectedViewFull() {
+        
         selectedLeadingConstraint.constant = 10
         selectedTrailingConstraint.constant = -10
         selectedView.layer.maskedCorners = [.layerMaxXMaxYCorner, .layerMaxXMinYCorner, .layerMinXMaxYCorner, .layerMinXMinYCorner]
